refactor: Rework 2FA and move unauthorized page
