+++ src/app/(root)/(common)/unauthorized/_components/unauthorized-card.tsx
@@ -8,7 +8,7 @@ export const UnauthorizedCard = () => {
       title="401 Unauthorized"
       description="You are not authorized to access this page."
       altActionHref="/"
-      altActionLinkText="Back home"
+      altActionLinkText="Back to home"
     >
       <LockKeyholeIcon className="mx-auto size-10 animate-shake text-destructive" />
     </AuthCard>
