+++ src/schemas/auth.ts
@@ -79,7 +79,6 @@ export const DeleteAccountSchema = z.object({
 });
 
 export const ToggleTwoFactorAuthenticationSchema = z.object({
-  isTwoFactorEnabled: z.boolean(),
   code: z.optional(z.string()),
 });
